Add seeds.

+++ hdcd-python/src/control.rs
@@ -37,7 +37,7 @@ pub fn control_from_pyobj(py: Python, obj: Option<PyObject>) -> PyResult<Control
         };
 
         if let Ok(pyvalue) = obj.getattr(py, "seed") {
-            if let Ok(value) = pyvalue.extract::<(usize,)>(py) {
+            if let Ok(value) = pyvalue.extract::<(u64,)>(py) {
                 control = control.with_seed(value.0);
             }
         };
